Extract timezone parsing in a dedicated module
